0.2.1+0.2: Introduced potential bug in receiver by making the global update happen on the socket response instead of constantly checking on update which needs testing. Fully replaced the old game object workflow for the new one. Replaced the update callback for one that passes over an object that contains info about the update. Added a toggle to the new manager for controlling init on start. Cleaned up the remaining uses of the var keyword. Deleted sender code to clean up project which will be reintroduced at a later point.
