--- SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/ClientSaveObject.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using SpeckleCore;
-
-namespace SpeckleUnity
-{
-	/// <summary>
-	/// 
-	/// </summary>
-	[Serializable]
-	public class ClientSaveObject
-	{
-		/// <summary>
-		/// 
-		/// </summary>
-		public SpeckleApiClient client;
-
-		/// <summary>
-		/// 
-		/// </summary>
-		public string key;
-	}
-}
