--- SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/ClientSaveObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e59eb789c049d16409eeea1d47c37048
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
