--- SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/SpeckleClient.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb7b7fbfc22a655419e0aa47dd25cdd2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
