--- SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/SpeckleReceiver.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d8da331d1797866409cc8c38229d85e9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
