--- SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/SpeckleSender.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2152b5111b5228145842c5eba8eb4d38
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
