+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/SpeckleUnityClient.cs
@@ -1,25 +1,20 @@
 ﻿using System;
+using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 using SpeckleCore;
 
-
 namespace SpeckleUnity
 {
-	/// <summary>
-	/// 
-	/// </summary>
 	[Serializable]
-	public abstract class SpeckleUnityClient : MonoBehaviour
+	public abstract class SpeckleUnityClient
 	{
-		/// <summary>
-		/// 
-		/// </summary>
-		protected string unityGUID;
+		protected SpeckleUnityManager controller;
 
 		/// <summary>
 		/// 
 		/// </summary>
-		public SpeckleApiClient client;
+		[HideInInspector] public SpeckleApiClient client;
 
 		/// <summary>
 		/// 
@@ -29,12 +24,7 @@ namespace SpeckleUnity
 		/// <summary>
 		/// 
 		/// </summary>
-		public string authToken = "";
-
-		/// <summary>
-		/// 
-		/// </summary>
-		protected virtual void AssignEvents ()
+		public virtual void RegisterClient ()
 		{
 			client.OnReady += ClientOnReady;
 			client.OnLogData += ClientOnLogData;
@@ -45,7 +35,7 @@ namespace SpeckleUnity
 		/// <summary>
 		/// 
 		/// </summary>
-		public virtual void DisposeClient ()
+		public virtual void UnregisterClient ()
 		{
 			client.OnReady -= ClientOnReady;
 			client.OnLogData -= ClientOnLogData;
@@ -58,24 +48,17 @@ namespace SpeckleUnity
 		/// <summary>
 		/// 
 		/// </summary>
-		/// <param name="URL"></param>
-		public abstract void InitializeClient (string URL);
-
-		/// <summary>
-		/// 
-		/// </summary>
-		/// <param name="client"></param>
-		public abstract void CompleteDeserialization (SpeckleApiClient client);
+		/// <param name="url"></param>
+		public abstract IEnumerator InitializeClient (SpeckleUnityManager controller, string url, string authToken);
 
 		/// <summary>
 		/// 
 		/// </summary>
 		/// <param name="source"></param>
 		/// <param name="e"></param>
-		public virtual void ClientOnReady (object source, SpeckleEventArgs e)
+		protected virtual void ClientOnReady (object source, SpeckleEventArgs e)
 		{
 			Debug.Log ("Client ready");
-			//Debug.Log(JsonConvert.SerializeObject(e.EventData));
 		}
 
 		/// <summary>
@@ -83,10 +66,9 @@ namespace SpeckleUnity
 		/// </summary>
 		/// <param name="source"></param>
 		/// <param name="e"></param>
-		public virtual void ClientOnLogData (object source, SpeckleEventArgs e)
+		protected virtual void ClientOnLogData (object source, SpeckleEventArgs e)
 		{
 			Debug.Log ("Client LogData");
-			//Debug.Log(JsonConvert.SerializeObject(e.EventData));
 		}
 
 		/// <summary>
@@ -94,10 +76,9 @@ namespace SpeckleUnity
 		/// </summary>
 		/// <param name="source"></param>
 		/// <param name="e"></param>
-		public virtual void ClientOnWsMessage (object source, SpeckleEventArgs e)
+		protected virtual void ClientOnWsMessage (object source, SpeckleEventArgs e)
 		{
 			Debug.Log ("Client WsMessage");
-			//Debug.Log(JsonConvert.SerializeObject(e.EventData));
 		}
 
 		/// <summary>
@@ -105,12 +86,9 @@ namespace SpeckleUnity
 		/// </summary>
 		/// <param name="source"></param>
 		/// <param name="e"></param>
-		public virtual void ClientOnError (object source, SpeckleEventArgs e)
+		protected virtual void ClientOnError (object source, SpeckleEventArgs e)
 		{
-			Debug.Log ("Client Error");
-			//Debug.Log(JsonConvert.SerializeObject(e.EventData));
+			Debug.LogError ("Client Error");
 		}
-
 	}
-
 }
