+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/SpeckleUnityClient.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9ced0de86511df0459b5ad63d6292d07
+guid: cb7b7fbfc22a655419e0aa47dd25cdd2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
