+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/SpeckleUnityReceiver.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bf521117ff37449499d286b1e2207cc1
+guid: d8da331d1797866409cc8c38229d85e9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
