+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/SpeckleUnitySender.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a228389880a067145a5c3f83b16e0acd
+guid: 2152b5111b5228145842c5eba8eb4d38
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
