+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/Web Sockets/SpeckleUnityUpdate.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 29c8e319c27dcc04d820766253747c89
+guid: 0d78bf04a66b48f418b6baa3fcbe2d4b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
