+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/Web Sockets/SpeckleUnityUpdateEvent.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Events;
+
+namespace SpeckleUnity
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	[Serializable]
+	public class SpeckleUnityUpdateEvent : UnityEvent<SpeckleUnityUpdate> { }
+}
