+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/Web Sockets/SpeckleUnityUpdateEvent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 20ba57babc4fe974da06bed2434666e3
+guid: cdb13527af84e4c49af19dd96388d738
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
