+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/Web Sockets/UpdateType.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SpeckleUnity
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public enum UpdateType
+    {
+        Global,
+        Meta,
+        Name,
+        Object,
+        Children
+    }
+}
