+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/Clients/Web Sockets/UpdateType.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f82fb3666afbd964a9c1de4262817917
+guid: 2a828f6bb16c7a84fbb993e55dbfc78f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
