--- SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/SendComponents.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3415ac22c5613324ebadea00e5399b85
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
