--- SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/SendComponents/SpeckleSendSlider.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db7a308e94d03584d9756fe7a00bf5f4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
