+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/SpeckleUnityManager.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Events;
 using SpeckleCore;
 using System.IO;
 using System.Runtime.Serialization.Formatters.Binary;
@@ -15,142 +16,63 @@ namespace SpeckleUnity
 		/// <summary>
 		/// 
 		/// </summary>
+		public bool initializeOnStart = true;
+
+		/// <summary>
+		/// 
+		/// </summary>
 		[Tooltip ("URL for the Speckle Server you want to use. eg: https://hestia.speckle.works/api/v1")]
-		[SerializeField] protected string ServerURL; 
+		public string serverUrl = "";
 
 		/// <summary>
 		/// 
 		/// </summary>
-		protected BinaryFormatter formatter;
+		public string authToken = "";
 
 		/// <summary>
 		/// 
 		/// </summary>
-		protected const string DATA_FILENAME = "TestSaveClient.dat";
+		public Material meshMaterial;
 
 		/// <summary>
 		/// 
 		/// </summary>
-		protected Dictionary<string, SpeckleApiClient> ClientSaveDictionary = new Dictionary<string, SpeckleApiClient> ();
+		public SpeckleUnityUpdateEvent onUpdateReceived;
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public SpeckleUnityReceiver[] receivers = Array.Empty<SpeckleUnityReceiver> ();
 
 		/// <summary>
 		/// 
 		/// </summary>
 		protected virtual void Start ()
 		{
+			if (initializeOnStart) Initialize ();
+		}
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public virtual void Initialize ()
+		{
 			SpeckleInitializer.Initialize ();
 			LocalContext.Init ();
-			formatter = new BinaryFormatter ();
 
-			LoadClients ();
-			SpeckleUnitySender[] SenderClients = FindObjectsOfType<SpeckleUnitySender> ();
-			foreach (var c in SenderClients)
+			for (int i = 0; i < receivers.Length; i++)
 			{
-				if (ClientSaveDictionary.ContainsKey (c.keyForSaving) && c.persistent)
-					c.CompleteDeserialization (ClientSaveDictionary[c.keyForSaving]);
-				else
-					c.InitializeClient (ServerURL);
+				StartCoroutine (receivers[i].InitializeClient (this, serverUrl, authToken));
 			}
-
-            SpeckleUnityReceiver[] ReceiverClients = FindObjectsOfType<SpeckleUnityReceiver> ();
-            foreach (var c in ReceiverClients)
-            {
-                c.InitializeClient (ServerURL);
-            }
-        }
-
-		/// <summary>
-		/// 
-		/// </summary>
-		protected virtual void OnApplicationQuit ()
-		{
-			Debug.Log ("Application ending after " + Time.time + " seconds");
-			SaveClients ();
 		}
 
-
-		/// <summary>
-		/// 
-		/// </summary>
-		protected virtual void SaveClients ()
+		// Update is called once per frame
+		/*protected virtual void Update ()
 		{
-			//check for persistent clients
-			SpeckleUnitySender[] Clients = FindObjectsOfType<SpeckleUnitySender> ();
-			foreach (var c in Clients)
+			for (int i = 0; i < receivers.Length; i++)
 			{
-				if (c.persistent)
-					ClientSaveDictionary[c.keyForSaving] = c.client;
-				else
-				{
-					ClientSaveDictionary.Remove (c.keyForSaving);
-					c.DisposeClient ();
-				}
+				receivers[i].OnWsMessageCheck ();
 			}
-			//TODO - dispose of clients in save file that no longer exist
-
-			//dictionaries aren't serializable?
-			List<ClientSaveObject> ClientSaveList = new List<ClientSaveObject> ();
-
-			foreach (var kvp in ClientSaveDictionary)
-			{
-				ClientSaveObject c = new ClientSaveObject
-				{
-					client = kvp.Value,
-					key = kvp.Key
-				};
-				ClientSaveList.Add (c);
-			}
-
-			try
-			{
-				// Create a FileStream that will write data to file.
-				FileStream writerFileStream =
-					new FileStream (DATA_FILENAME, FileMode.Create, FileAccess.Write);
-				// Save information
-				this.formatter.Serialize (writerFileStream, ClientSaveList);
-
-				// Close the writerFileStream when we are done.
-				writerFileStream.Close ();
-			}
-			catch (Exception)
-			{
-				Debug.Log ("Unable to save");
-			} // end try-catch
-		}
-
-		/// <summary>
-		/// 
-		/// </summary>
-		protected virtual void LoadClients ()
-		{
-			// Check if we had previously Save information 
-			if (File.Exists (DATA_FILENAME))
-			{
-				try
-				{
-					// Create a FileStream will gain read access to the data file
-					FileStream readerFileStream = new FileStream (DATA_FILENAME, FileMode.Open, FileAccess.Read);
-					// Reconstruct information
-					List<ClientSaveObject> ClientSaveList = (List<ClientSaveObject>)this.formatter.Deserialize (readerFileStream);
-
-					// Close the readerFileStream when we are done
-					readerFileStream.Close ();
-
-					//rebuild dictionary for searching later
-					foreach (var c in ClientSaveList)
-					{
-						ClientSaveDictionary[c.key] = c.client;
-					}
-
-				}
-				catch (Exception e)
-				{
-					Debug.Log ("Could not read file: " + e.ToString ());
-				} // end try-catch
-
-			} // end if
-
-		}
-
+		}*/
 	}
 }
