+++ SpeckleUnity/Assets/SpeckleUnity/Runtime/Scripts/SpeckleUnityManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2a5f374aa5217d74ca279457483027d7
+guid: 29c8e319c27dcc04d820766253747c89
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
