+++ SpeckleUnity/Assets/SpeckleUnity/package.json
@@ -2,7 +2,7 @@
   "name": "com.open.speckleunity",
   "displayName": "SpeckleUnity",
   "author": "Christopher Morse, Jak Hussain, Dean Giddy, Luis E. Fraguadam",
-  "version": "0.2.1-preview",
+  "version": "0.3.0-preview",
   "unity": "2019.3",
   "description": "Send and receive Speckle Streams from your Unity application to a Speckle Server of your choice. \n\nRequires .NET 4.x and does not support IL2CPP platforms.",
   "keywords": [
